refactor: move slider components to ui slide folder
